Memory cards manager: bugfixes (all of them when multitap 1 is disabled) + better internal terminology: Terminology: Port (used only for presentation): actual PS2 port: 1-based index  (1 or 2 or multitap X port Y, whrere X is 1/2 and Y is 1/2/3/4 ) list-view-index: the running index (0-based) on the MCD viewable list (first always 0, last is always <num-items-on-the-list>-1). Slot: pcsx2 internal slot representation. the slots are always fixed as follows: slot 0: port 1 (or multitap 1 port 1 if multitap 1 is enabled) slot 1: port 2 (or multitap 2 port 1 if multitap 2 is enabled) slots 2,3,4: multitap 1: remaining 3 ports slots 5,6,7: multitap 2: remaining 3 ports
The terms view-index and slot were used interchangeably (and always called 'slot'), and this was causing many bugs when multitap 1 is disabled (because on that specific case the view index doesn't align with the slot - because the multitap 1 items are not on the list). It should now be easier to follow the code and further modify it.

git-svn-id: https://pcsx2.googlecode.com/svn/trunk@4410 96395faa-99c1-11dd-bbfe-3dabce05a288