GSdx: CRC Hacks: Import from Cutie - PLEASE HELP WITH TESTING: - 14 Updated hacks, 41 new hacks (See full list at the first comment). - New crc-hacks might have patial crc lists, so it's possible that some copies won't get the benefit (yet). - Non crc-hacks code was NOT imported from Cutie, so some crc hacks might not be as effective as within Cutie itself. - New region: CH, few more CRCs.
Note:
Due to the very big list of affected games, I couldn't test the vast majority of them, and so your help would be appreciated in testing. Please report (even if someone else already reported with the same result as yours) with the following format:
<game name> - <better/worse/same> (reason) - tested <little/medium/much>
e.g.:
Tomb Raider Legend - Better (removes garbage graphics) - tested a little.
GTA San Andreas - Better (removes ghosts) - tested a little.

And of course, big thank you goes to the author of Cutie, for the time put into improving PCSX2, and for respecting the GPL license and releasing the code.

git-svn-id: https://pcsx2.googlecode.com/svn/trunk@5214 96395faa-99c1-11dd-bbfe-3dabce05a288